// src/app/interfaces/orderModel.ts
export interface IOrderModel {
    NBOXCOUNT: number;
    NBOXID: number;
    DTDROPDATE: string;
    DTPICKUPDATE: string;
    NDROPLAT: number;
    NDROPLONG: number;
    NPICKUPLAT: number;
    NPICKUPLONG: number;
    NDISTANCE: number;
    NCLIENTID: number;
    SADDRESSPICKUP: string;
    SADDRESSDROP: string;
  }